Normalize usage of StrBoolAttr
Priori to this patch, StrBoolAttr attributes had the strange property to be
either unset, set to false or set to true. And the associated value was a string
containing either "true" or "false". This is both cumbersome to use and
inefficient.

This patch propose to no longer use the value filed, and just check the presence
of the attribute, which was already an established practice in part of the LLVM
codebase, see for instance https://reviews.llvm.org/D96400
As a side effect, when compiling sqlite3.c source code in -O0, this change makes
the instruction count decrease from 6,157,211,048 to 6,155,407,798. Simpler and
faster.

This patch also provides an auto-upgrade path to ensure the normalization
actually happens, and asserts if someone tries to use the API in an inconsistent
way.

Differential Revision: https://reviews.llvm.org/D99080